separated the implementation of the homescreen drawers to separate files
