+++ threddit_app/lib/features/home_page/view/screens/home_screen.dart
@@ -1,14 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:threddit_clone/app/route.dart';
-import 'package:threddit_clone/features/home_page/view/widgets/communities_tiles.dart';
-import 'package:threddit_clone/features/home_page/view/widgets/following_tiles.dart';
+import 'package:threddit_clone/features/home_page/view/widgets/left_drawer.dart';
 import 'package:threddit_clone/features/home_page/view/widgets/right_drawer.dart';
 import 'package:threddit_clone/features/listing/view/widgets/feed_widget.dart';
-import 'package:threddit_clone/features/user_system/view_model/sign_in_with_google/google_auth_controller.dart';
 import 'package:threddit_clone/theme/colors.dart';
-import 'package:threddit_clone/theme/photos.dart';
 import 'package:threddit_clone/theme/text_styles.dart';
 
 class HomeScreen extends ConsumerStatefulWidget {
@@ -28,12 +24,11 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
 
   @override
   Widget build(BuildContext context) {
-    ////final user = ref.watch(userProvider)!;
     return Scaffold(
       key: _scaffoldKey,
       backgroundColor: AppColors.backgroundColor,
       appBar: AppBar(
-        //automaticallyImplyLeading: false,
+        automaticallyImplyLeading: false,
         iconTheme: const IconThemeData(color: Colors.white),
         actions: [
           IconButton(
@@ -55,9 +50,7 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
           ),
           textStyle: AppTextStyles.primaryTextStyle,
           dropdownMenuEntries:
-
-              ///use provider here
-              tabs.map<DropdownMenuEntry<String>>((String string) {
+            tabs.map<DropdownMenuEntry<String>>((String string) {
             return DropdownMenuEntry(value: string, label: string);
           }).toList(),
           width: 150,
@@ -66,7 +59,6 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
             setState(() {
               feedID = value!;
             });
-
             ScaffoldMessenger.of(context).showSnackBar(SnackBar(
               content: Text('Changed to tab $value'),
               duration: Durations.short1,
@@ -74,111 +66,8 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
           },
         ),
       ),
-      drawer: Theme(
-        data: Theme.of(context).copyWith(canvasColor: AppColors.mainColor),
-        child: Drawer(
-          elevation: double.maxFinite,
-          backgroundColor: AppColors.mainColor,
-          shadowColor: AppColors.mainColor,
-          surfaceTintColor: AppColors.mainColor,
-          child: ListView(
-            padding: EdgeInsets.zero,
-            children: <Widget>[
-              DrawerHeader(
-                child: Text(
-                  'Recently Visited',
-                  style: AppTextStyles.primaryTextStyle
-                      .copyWith(fontWeight: FontWeight.bold),
-                ),
-              ),
-              const CommunitiesTiles(title: "Communities"),
-              const FollowingTiles(title: "Following"),
-            ],
-          ),
-        ),
-      ),
-      endDrawer: Drawer(
-        backgroundColor: AppColors.mainColor,
-        child: Column(
-          children: [
-            SizedBox(
-              height: 150.h,
-              child: DrawerHeader(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    InkWell(
-                      onTap: () {
-                        Navigator.pushNamed(
-                            context, RouteClass.userProfileScreen);
-                      },
-                      child: Image.asset(
-                        Photos.snoLogo,
-                        width: 50.w,
-                        height: 50.h,
-                      ),
-                    ),
-                    Text(
-                      'u/UserName',
-                      style: AppTextStyles.primaryTextStyle,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            RightDrawerButtons(
-                icon: const Icon(
-                  Icons.person_outline,
-                  color: AppColors.whiteColor,
-                ),
-                title: "My profile",
-                onTap: () {
-                  Navigator.pushNamed(context, RouteClass.userProfileScreen);
-                }),
-            RightDrawerButtons(
-                icon: const Icon(
-                  Icons.group_add_outlined,
-                  color: AppColors.whiteColor,
-                ),
-                title: "Create a community",
-                onTap: () {
-                  Navigator.pushNamed(
-                      context, RouteClass.createCommunityScreen);
-                }),
-            RightDrawerButtons(
-                icon: const Icon(
-                  Icons.bookmarks_outlined,
-                  color: AppColors.whiteColor,
-                ),
-                title: "Saved",
-                onTap: () {}),
-            RightDrawerButtons(
-                icon: const Icon(
-                  Icons.history_toggle_off_rounded,
-                  color: AppColors.whiteColor,
-                ),
-                title: "History",
-                onTap: () {}),
-            RightDrawerButtons(
-              icon: const Icon(
-                Icons.logout_rounded,
-                color: AppColors.whiteColor,
-              ),
-              title: "Logout",
-              onTap: () => ref.watch(authControllerProvider.notifier).logout(),
-            ),
-            RightDrawerButtons(
-                icon: const Icon(
-                  Icons.settings_outlined,
-                  color: AppColors.whiteColor,
-                ),
-                title: "Settings",
-                onTap: () {
-                  Navigator.pushNamed(context, RouteClass.accountSettingScreen);
-                }),
-          ],
-        ),
-      ),
+      drawer: const LeftDrawer(),
+      endDrawer: const RightDrawer(),
       body: FeedWidget(feedID: feedID),
     );
   }
